# src/api-bma/api-bma/bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "api-bma, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/api/Beneficiarios": {
      "get": {
        "tags": [
          "Beneficiarios"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Beneficiarios"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Beneficiario"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Beneficiarios/{id}": {
      "get": {
        "tags": [
          "Beneficiarios"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Beneficiarios"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Beneficiario"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Beneficiarios"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "definitions": {
    "Beneficiario": {
      "type": "object",
      "properties": {
        "links": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/LinkDto"
          }
        },
        "beneficiarioId": {
          "format": "int32",
          "type": "integer"
        },
        "titularBolsaFamilia": {
          "$ref": "#/definitions/Titularbolsafamilia"
        },
        "quantidadeDependentes": {
          "format": "int32",
          "type": "integer"
        },
        "dependentes": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Dependente"
          }
        },
        "endereco": {
          "$ref": "#/definitions/Endereco"
        }
      }
    },
    "Dependente": {
      "required": [
        "data_Nascimento",
        "nome"
      ],
      "type": "object",
      "properties": {
        "links": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/LinkDto"
          }
        },
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "nome": {
          "type": "string"
        },
        "data_Nascimento": {
          "format": "date",
          "type": "string"
        },
        "beneficiarioId": {
          "format": "int32",
          "type": "integer"
        },
        "beneficiario": {
          "$ref": "#/definitions/Beneficiario"
        }
      }
    },
    "Endereco": {
      "type": "object",
      "properties": {
        "links": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/LinkDto"
          }
        },
        "enderecoId": {
          "format": "int32",
          "type": "integer"
        },
        "cep": {
          "type": "string"
        },
        "state": {
          "type": "string"
        },
        "city": {
          "type": "string"
        },
        "neighborhood": {
          "type": "string"
        },
        "street": {
          "type": "string"
        },
        "beneficiarioId": {
          "format": "int32",
          "type": "integer"
        },
        "beneficiario": {
          "$ref": "#/definitions/Beneficiario"
        }
      }
    },
    "LinkDto": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "href": {
          "type": "string"
        },
        "rel": {
          "type": "string"
        },
        "metodo": {
          "type": "string"
        }
      }
    },
    "Titularbolsafamilia": {
      "type": "object",
      "properties": {
        "nis": {
          "type": "string"
        },
        "nome": {
          "type": "string"
        }
      }
    }
  }
}